requirements.txt
google-generativeai==0.8.4
python-dotenv==1.1.0
tqdm==4.67.1
grpcio==1.71.0
